+++ js/my.js
@@ -1,13 +1,15 @@
 // hi quality variants
 function hq(kid, mom, dad) {
-    return (kid.DP > 7 && mom.DP > 7 && dad.DP > 7 
-         && mom.GQ > 10 && dad.GQ > 10 && kid.GQ > 10)
+    return (kid.DP > 6 && mom.DP > 6 && dad.DP > 6 && mom.GQ > 9 && dad.GQ > 9 && kid.GQ > 9)
+}
+
+function hq1(sample) {
+  return !sample.unknown && sample.DP > 6 && sample.GQ > 9
 }
 
 function hqrv(variant, INFO, af_cutoff) {
-	// hi-quality, rare variant.
-	return (!('gnomad_popmax_af_filter' in INFO) && variant.FILTER == 'PASS'
-         && (!variant.is_multiallelic) && INFO.gnomad_popmax_af < af_cutoff)
+  // hi-quality, rare variant.
+  return variant.FILTER == 'PASS' && INFO.gnomad_popmax_af < af_cutoff
 }
 
 function denovo(kid, mom, dad){
@@ -55,16 +57,12 @@ function x_recessive(kid, mom, dad) {
               && kid.sex == 'male' && mom.AB > 0.25 && mom.AB < 0.75 && kid.AB > 0.75)
 }
 
-function lenient_denovo(kid, mom, dad){
-  if(!hq(kid, mom, dad)){ return false; }
-  // check genotypes match expected pattern
-  if(!(kid.alts == 1 && mom.alts == 0 && dad.alts == 0)){ return false; }
-  return ((mom.AD[1] + dad.AD[1]) < 2) && kid.AB > 0.2 && kid.AB < 0.8;
+
+// heterozygous (1 side of compound het)
+function solo_ch_het_side(sample) {
+  return hq1(sample) && sample.AB > 0.2 && sample.AB < 0.8
 }
 
-function lenient_ar(kid, mom, dad) {
-	 return hq(kid, mom, dad) &&
-       (mom.alts == 1 || dad.alts == 1) && kid.alts >= 1 && mom.alts != -1 && dad.alts != -1 &&
-        dad.alts != 2 && mom.alts != 2
-        && (kid.AB > 0.2 && kid.AB < 0.8)
+function comphet_side(kid, mom, dad) {
+  return solo_ch_het_side(kid) && hq1(mom) && hq1(dad) && (solo_ch_het_side(mom) != solo_ch_het_side(dad)) && mom.alts != 2 && dad.alts != 2;
 }